move constants.js to project root; update relative imports and package.json entry point
