+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "airswap-maker-kit",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "handlers.js",
   "repository": "git@github.com:airswap/airswap-maker-kit.git",
   "author": "Don Mosites <don.mosites@fluidity.io>",
   "license": "MIT",
