+++ scripts/indexer/enable.js
@@ -1,7 +1,7 @@
 const ethers = require('ethers')
 const network = require('../lib/network.js')
 const prompt = require('../lib/prompt.js')
-const constants = require('../constants.js')
+const constants = require('../../constants.js')
 
 const IERC20 = require('@airswap/tokens/build/contracts/IERC20.json')
 const indexerDeploys = require('@airswap/indexer/deploys.json')
